2023.11.02. 회원가입 파일업로드 진행중

+++ Project/src/main/java/com/project/vodto/UploadFile.java
@@ -17,5 +17,5 @@ public class UploadFile {
 	private String extension;
 	private String originalFileName;
 	private String newFileName;
-	private int fileSize;
+	private long fileSize;
 }
